+++ src/encode/responsestatus.go
@@ -49,6 +49,7 @@ var ResponseMessage = map[ResStatus]int{
 	ErrClusterConnect:  921,
 	ErrClusterNotfound: 922,
 	ErrClusterParams:   923,
+	ErrClusterList:     924,
 
 	ErrNamespaceNotfound: 940,
 	ErrNameNotfound:      941,
@@ -65,6 +66,10 @@ var ResponseMessage = map[ResStatus]int{
 	ErrSecretMarshal:   1120,
 	ErrSecretImageSave: 1121,
 	ErrSecretDelete:    1122,
+
+	ErrNodeCordon:   1140,
+	ErrNodeNotfound: 1141,
+	ErrNodeDrain:    1142,
 }
 
 const (
@@ -86,6 +91,7 @@ const (
 	ErrClusterConnect  ResStatus = "集群链接错误"
 	ErrClusterNotfound ResStatus = "集群不存在"
 	ErrClusterParams   ResStatus = "集群参数错误"
+	ErrClusterList     ResStatus = "集群列表错误"
 
 	ErrSecretMarshal   ResStatus = "转换错误"
 	ErrSecretImageSave ResStatus = "添加镜像Secret错误"
@@ -121,6 +127,10 @@ const (
 	ErrSysRoleUserLen      ResStatus = "请选择用户"
 	ErrSysRoleUserDelete   ResStatus = "角色删除失败"
 	ErrSysUserNotfound     ResStatus = "用户不存在"
+
+	ErrNodeCordon   ResStatus = "操作失败"
+	ErrNodeNotfound ResStatus = "节点不存在"
+	ErrNodeDrain    ResStatus = "驱逐失败"
 )
 
 func (c ResStatus) String() string {
