+++ src/pkg/cronjob/http.go
@@ -5,19 +5,17 @@
  * @Software: GoLand
  */
 
-package nodes
+package cronjob
 
 import (
 	"context"
 	"encoding/json"
-	"github.com/pkg/errors"
-	"net/http"
-	"strconv"
-
 	valid "github.com/asaskevich/govalidator"
 	"github.com/go-kit/kit/endpoint"
 	kithttp "github.com/go-kit/kit/transport/http"
 	"github.com/gorilla/mux"
+	"github.com/pkg/errors"
+	"net/http"
 
 	"github.com/kplcloud/kplcloud/src/encode"
 )
@@ -29,59 +27,18 @@ func MakeHTTPHandler(s Service, dmw []endpoint.Middleware, opts []kithttp.Server
 
 	eps := NewEndpoint(s, map[string][]endpoint.Middleware{
 		"Sync": ems,
-		"Info": ems,
 	})
 
 	r := mux.NewRouter()
 
-	r.Handle("/{cluster}/sync", kithttp.NewServer(
+	r.Handle("/{cluster}/sync/{namespace}", kithttp.NewServer(
 		eps.SyncEndpoint,
 		kithttp.NopRequestDecoder,
 		encode.JsonResponse,
 		opts...,
 	)).Methods(http.MethodGet)
-	r.Handle("/{cluster}/list", kithttp.NewServer(
-		eps.ListEndpoint,
-		decodeListRequest,
-		encode.JsonResponse,
-		opts...,
-	)).Methods(http.MethodGet)
-	r.Handle("/{cluster}/info/{name}", kithttp.NewServer(
-		eps.InfoEndpoint,
-		decodeInfoRequest,
-		encode.JsonResponse,
-		opts...,
-	)).Methods(http.MethodGet)
 
 	return r
-}
-
-func decodeInfoRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	var req infoRequest
-
-	vars := mux.Vars(r)
-	name, ok := vars["name"]
-	if !ok {
-		return nil, encode.InvalidParams.Error()
-	}
-	req.Name = name
-
-	return req, nil
-}
-
-func decodeListRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	var req listRequest
-	page, _ := strconv.Atoi(r.URL.Query().Get("page"))
-	pageSize, _ := strconv.Atoi(r.URL.Query().Get("pageSize"))
-	if page <= 0 {
-		page = 1
-	}
-	if pageSize <= 1 {
-		pageSize = 10
-	}
-	req.page = page
-	req.pageSize = pageSize
-	return req, nil
 }
 
 func decodeSyncRequest(_ context.Context, r *http.Request) (interface{}, error) {
