+++ src/pkg/nodes/endpoint.go
@@ -28,22 +28,66 @@ type (
 	listRequest struct {
 		page, pageSize int
 	}
+
+	infoRequest struct {
+		Name string
+	}
+	infoResult struct {
+		Name             string            `json:"name"`
+		Status           string            `json:"status"`
+		CPU              string            `json:"cpu"`
+		Memory           string            `json:"memory"`
+		UsedCPU          string            `json:"usedCpu"`
+		UsedMemory       string            `json:"usedMemory"`
+		Bandwidth        string            `json:"bandwidth"`  // 带宽
+		SystemDisk       string            `json:"systemDisk"` // 系统盘大小
+		InternalIp       string            `json:"internalIp"`
+		ExternalIp       string            `json:"externalIp"`
+		KubeletVersion   string            `json:"kubeletVersion"`
+		KubeProxyVersion string            `json:"kubeProxyVersion"`
+		ContainerVersion string            `json:"containerVersion"`
+		OsImage          string            `json:"osImage"`
+		Scheduled        bool              `json:"scheduled"`
+		Remark           string            `json:"remark"`
+		Labels           map[string]string `json:"labels"`
+		PodNum           int64             `json:"podNum"`
+	}
 )
 
 type Endpoints struct {
 	SyncEndpoint endpoint.Endpoint
 	ListEndpoint endpoint.Endpoint
+	InfoEndpoint endpoint.Endpoint
 }
 
 func NewEndpoint(s Service, dmw map[string][]endpoint.Middleware) Endpoints {
 	eps := Endpoints{
 		SyncEndpoint: makeSyncEndpoint(s),
+		InfoEndpoint: makeInfoEndpoint(s),
 	}
 
 	for _, m := range dmw["Sync"] {
 		eps.SyncEndpoint = m(eps.SyncEndpoint)
 	}
+	for _, m := range dmw["Info"] {
+		eps.InfoEndpoint = m(eps.InfoEndpoint)
+	}
 	return eps
+}
+
+func makeInfoEndpoint(s Service) endpoint.Endpoint {
+	return func(ctx context.Context, request interface{}) (response interface{}, err error) {
+		clusterId, ok := ctx.Value(middleware.ContextKeyClusterId).(int64)
+		if !ok {
+			return nil, encode.ErrClusterNotfound.Error()
+		}
+		req := request.(infoRequest)
+		res, err := s.Info(ctx, clusterId, req.Name)
+		return encode.Response{
+			Data:  res,
+			Error: err,
+		}, err
+	}
 }
 
 func makeSyncEndpoint(s Service) endpoint.Endpoint {
