+++ src/repository/cluster/cache.go
@@ -28,6 +28,10 @@ type cache struct {
 	kitcache kitcache.Service
 }
 
+func (s *cache) List(ctx context.Context, name string, status int, page, pageSize int) (res []types.Cluster, total int, err error) {
+	return s.next.List(ctx, name, status, page, pageSize)
+}
+
 func (s *cache) SaveRole(ctx context.Context, clusterRole *types.ClusterRole, roles []types.PolicyRule) (err error) {
 	return s.next.SaveRole(ctx, clusterRole, roles)
 }
