+++ src/repository/types/cluster.go
@@ -15,8 +15,9 @@ type Cluster struct {
 	Alias      string     `gorm:"column:alias;comment:'别名'" json:"alias"`                       // 别名
 	Remark     string     `gorm:"column:remark;comment:'备注'" json:"remark"`                     // 备注
 	Label      string     `gorm:"column:label;comment:'标签'" json:"label"`                       // 标签
-	Status     int        `gorm:"column:status;default:0;comment:'状态:0未启用,1:正常'" json:"status"` // 状态
+	Status     int        `gorm:"column:status;default:0;comment:'状态:1未启用,2:正常'" json:"status"` // 状态
 	ConfigData string     `gorm:"column:config_data;type:text;comment:'配置文件'" json:"config_data"`
+	NodeNum    int        `gorm:"column:node_num;comment:'节点数量'" json:"node_num"`
 	CreatedAt  time.Time  `gorm:"column:created_at" json:"created_at"` // 创建时间
 	UpdatedAt  time.Time  `gorm:"column:updated_at" json:"updated_at"` // 更新时间
 	DeletedAt  *time.Time `gorm:"column:deleted_at" json:"deleted_at"`
