+++ src/repository/types/project.go
@@ -1,30 +1,24 @@
-/**
- * Created by GoLand.
- * User: zghua
- * Email: xzghua@gmail.com
- * Date: 2019-06-25
- * Time: 17:10
- */
+// Package types
 package types
 
 import "time"
 
 type Project struct {
-	Id          int64  `json:"id"`
-	Alias       string `json:"alias"`
-	Name        string `json:"name"`
-	Namespace   string `json:"namespace"`
-	Cpu         int64  `json:"cpu"`
-	MaxCpu      int64  `json:"max_cpu"`
-	Memory      int64  `json:"memory"`
-	MaxMemory   int64  `json:"max_memory"`
-	GitRepo     string `json:"git_repo"`
-	Version     string `json:"version"`
-	Status      string `json:"status"`
-	State       int    `json:"state"`
-	Remark      string `json:"remark"`
-	AuditStatus int    `json:"audit_status"`
-	Step        int    `json:"step"`
+	Id          int64  `gorm:"column:id;primary_key" json:"id"`
+	Alias       string `gorm:"column:alias;notnull;comment:'别名'" json:"alias"`
+	Name        string `gorm:"column:name;index;24;notnull;comment:'名称'" json:"name"`
+	Namespace   string `gorm:"column:namespace;index;24;notnull;;comment:'空间'" json:"namespace"`
+	Cpu         int64  `gorm:"column:cpu;null;comment:'基础CPU'" json:"cpu"`
+	MaxCpu      int64  `gorm:"column:max_cpu;null;comment:'最大CPU'" json:"max_cpu"`
+	Memory      int64  `gorm:"column:memory;null;comment:'基础内存'" json:"memory"`
+	MaxMemory   int64  `gorm:"column:max_memory;null;comment:'最大内存'" json:"max_memory"`
+	GitRepo     string `gorm:"column:git_repo;null;comment:'git仓库'" json:"git_repo"`
+	Version     string `gorm:"column:alias;notnull;comment:'别名'" json:"version"`
+	Status      string `gorm:"column:alias;notnull;comment:'别名'" json:"status"`
+	State       int    `gorm:"column:alias;notnull;comment:'别名'" json:"state"`
+	Remark      string `gorm:"column:alias;notnull;comment:'别名'" json:"remark"`
+	AuditStatus int    `gorm:"column:alias;notnull;comment:'别名'" json:"audit_status"`
+	Step        int    `gorm:"column:alias;notnull;comment:'别名'" json:"step"`
 
 	CreatedAt time.Time  `gorm:"column:created_at" json:"created_at" form:"created_at"` // 创建时间
 	UpdatedAt time.Time  `gorm:"column:updated_at" json:"updated_at" form:"updated_at"` // 更新时间
